+++ tests/test_raillabel/format/test_frame_interval.py
@@ -62,6 +62,44 @@ def test_len():
     assert len(frame_interval) == 5
 
 
+def test_from_frame_uids_empty():
+    frame_uids = []
+
+    assert FrameInterval.from_frame_uids(frame_uids) == []
+
+def test_from_frame_uids_one_frame():
+    frame_uids = [1]
+
+    assert FrameInterval.from_frame_uids(frame_uids) == [
+        FrameInterval(1, 1)
+    ]
+
+def test_from_frame_uids_one_interval():
+    frame_uids = [1, 2, 3, 4]
+
+    assert FrameInterval.from_frame_uids(frame_uids) == [
+        FrameInterval(1, 4)
+    ]
+
+def test_from_frame_uids_multiple_intervals():
+    frame_uids = [0, 1, 2, 3, 6, 7, 9, 12, 13, 14]
+
+    assert FrameInterval.from_frame_uids(frame_uids) == [
+        FrameInterval(0, 3),
+        FrameInterval(6, 7),
+        FrameInterval(9, 9),
+        FrameInterval(12, 14),
+    ]
+
+def test_from_frame_uids_unsorted():
+    frame_uids = [5, 2, 1, 3]
+
+    assert FrameInterval.from_frame_uids(frame_uids) == [
+        FrameInterval(1, 3),
+        FrameInterval(5, 5),
+    ]
+
+
 if __name__ == "__main__":
     os.system("clear")
     pytest.main([__file__, "--disable-pytest-warnings", "--cache-clear", "-v"])